remove cache objects usage from todebug and use a model instead.
cache objects takes a couple minutes to run on my database and this
code will block the main thread until it finishes, then clears it and
calls for a refresh. new code simply runs a sql query to get at just
the objects we want using toeventquery and is much faster (and doesn't
block).



git-svn-id: https://svn.code.sf.net/p/tora/code/trunk@2948 4a83093e-0cc6-42a4-b7cd-3300864c03d0