Workaround for ORA 3127.
For some unknown reason it is ilegal to work with LONGs and LOBs
on one connection in parallel. OCI is re-entrant but sometimes 
OCI call end up with the error:
ORA-3127 no new operations allowed until the active operation ends

This is a temporary workaround that wraps two OCI calls(OCILobOpen, OCILobClose)
into while loop. Maybe in the future it will be mandatory to wrapp all the OCI calls.

NOTE: this patch prevents Tora from beeing compile-able on MSVC, there is no usleep
in MSVC.



git-svn-id: https://svn.code.sf.net/p/tora/code/trunk@3876 4a83093e-0cc6-42a4-b7cd-3300864c03d0